refactor(channels): centralize turn execution lifecycle
