+++ src/channels/turn/kernel.ts
@@ -1,38 +1,14 @@
-import type { ReplyPayload } from "../../auto-reply/reply-payload.js";
-import {
-  clearChannelHistoryIfEnabled,
-  recordChannelHistoryEntryWithMedia,
-} from "../../auto-reply/reply/history.js";
-import { dispatchReplyWithBufferedBlockDispatcher } from "../../auto-reply/reply/provider-dispatcher.js";
-import { runWithSessionInitConflictRetry } from "../../auto-reply/reply/session-init-conflict-retry.js";
-import type { FinalizedMsgContext } from "../../auto-reply/templating.js";
-import { resolveStorePath } from "../../config/sessions/paths.js";
-import {
-  createDiagnosticTraceContextFromActiveScope,
-  runWithDiagnosticTraceContext,
-} from "../../infra/diagnostic-trace-context.js";
-import { createSubsystemLogger } from "../../logging/subsystem.js";
+import { recordChannelHistoryEntryWithMedia } from "../../auto-reply/reply/history.js";
 import { toHistoryMediaEntries } from "../inbound-event/media.js";
-import { createChannelReplyPipeline } from "../message/reply-pipeline.js";
-import { recordInboundSession } from "../session.js";
-import { recordChannelBotPairLoopAndCheckSuppression } from "./bot-loop-protection.js";
 import {
-  EMPTY_CHANNEL_TURN_DISPATCH_COUNTS,
-  hasVisibleChannelTurnDispatch,
-  type ChannelTurnDispatchResultLike,
-  type ChannelTurnVisibleDeliverySignals,
-} from "./dispatch-result.js";
-import {
-  deliverInboundReplyWithMessageSendContext,
-  isDurableInboundReplyDeliveryHandled,
-  throwIfDurableInboundReplyDeliveryFailed,
-} from "./durable-delivery.js";
+  assembleResolvedChannelTurn,
+  dispatchAssembledChannelTurn,
+  runPreparedInboundReply,
+} from "./lifecycle.js";
 
 export { recordChannelBotPairLoopAndCheckSuppression } from "./bot-loop-protection.js";
 
 export type { ChannelBotLoopProtectionFacts } from "./bot-loop-protection.js";
-
-const NO_ADDITIONAL_DELIVERY_SIGNALS: ChannelTurnVisibleDeliverySignals = {};
 
 export {
   deliverInboundReplyWithMessageSendContext,
@@ -44,18 +20,13 @@ export type {
   DurableInboundReplyDeliveryParams,
 } from "./durable-delivery.js";
 import type {
-  AssembledChannelTurn,
-  ChannelTurnPlan,
   ChannelEventClass,
   ChannelTurnAdmission,
   ChannelEventDeliveryAdapter,
-  ChannelTurnHistoryFinalizeOptions,
   ChannelTurnLogEvent,
   ChannelTurnResult,
-  ChannelTurnResolved,
   DispatchedChannelTurnResult,
   NormalizedTurnInput,
-  PreparedChannelTurn,
   PreflightFacts,
   RunChannelTurnParams,
 } from "./types.js";
@@ -66,39 +37,17 @@ export {
   resolveChannelTurnDispatchCounts,
 } from "./dispatch-result.js";
 export type { ChannelTurnResult, DispatchedChannelTurnResult } from "./types.js";
+export {
+  dispatchAssembledChannelTurn,
+  dispatchChannelInboundReply,
+  dispatchChannelInboundTurn,
+  runPreparedInboundReply,
+} from "./lifecycle.js";
 
 const DEFAULT_EVENT_CLASS: ChannelEventClass = {
   kind: "message",
   canStartAgentTurn: true,
 };
-const log = createSubsystemLogger("channels/turn/kernel");
-
-function assembleResolvedChannelTurn<TDispatchResult>(
-  value: ChannelTurnResolved<TDispatchResult>,
-): AssembledChannelTurn | PreparedChannelTurn<TDispatchResult> {
-  if (!("route" in value)) {
-    return value;
-  }
-  if ("runDispatch" in value) {
-    const { cfg, route, ...turn } = value;
-    return {
-      ...turn,
-      routeSessionKey: route.sessionKey,
-      storePath: resolveStorePath(cfg.session?.store, { agentId: route.agentId }),
-      recordInboundSession,
-    };
-  }
-  const { cfg, route, ...turn } = value;
-  return {
-    ...turn,
-    cfg,
-    agentId: route.agentId,
-    routeSessionKey: route.sessionKey,
-    storePath: resolveStorePath(cfg.session?.store, { agentId: route.agentId }),
-    recordInboundSession,
-    dispatchReplyWithBufferedBlockDispatcher,
-  };
-}
 
 function isAdmission(value: unknown): value is ChannelTurnAdmission {
   if (!value || typeof value !== "object") {
@@ -140,17 +89,6 @@ function createNoopChannelEventDeliveryAdapter(): ChannelEventDeliveryAdapter {
       visibleReplySent: false,
     }),
   };
-}
-
-function clearPendingHistoryAfterTurn(params?: ChannelTurnHistoryFinalizeOptions): void {
-  if (!params?.isGroup || !params.historyKey || !params.historyMap || params.limit === undefined) {
-    return;
-  }
-  clearChannelHistoryIfEnabled({
-    historyMap: params.historyMap,
-    historyKey: params.historyKey,
-    limit: params.limit,
-  });
 }
 
 function resolveDroppedHistorySender(input: NormalizedTurnInput, preflight: PreflightFacts) {
@@ -217,469 +155,6 @@ async function recordDroppedChannelTurnHistory(params: {
 }
 
 export const recordDroppedChannelInboundHistory = recordDroppedChannelTurnHistory;
-
-function resolveAssembledReplyPipeline(
-  params: AssembledChannelTurn,
-): Pick<AssembledChannelTurn, "dispatcherOptions" | "replyOptions"> {
-  const turnAdoptionLifecycle =
-    params.turnAdoptionLifecycle ?? params.replyOptions?.turnAdoptionLifecycle;
-  if (!params.replyPipeline) {
-    return {
-      dispatcherOptions: params.dispatcherOptions,
-      replyOptions: turnAdoptionLifecycle
-        ? { ...params.replyOptions, turnAdoptionLifecycle }
-        : params.replyOptions,
-    };
-  }
-  const { onModelSelected, ...replyPipeline } = createChannelReplyPipeline({
-    cfg: params.cfg,
-    agentId: params.agentId,
-    channel: params.channel,
-    accountId: params.accountId,
-    ...params.replyPipeline,
-  });
-  return {
-    dispatcherOptions: {
-      ...replyPipeline,
-      ...params.dispatcherOptions,
-    },
-    replyOptions: {
-      onModelSelected,
-      ...params.replyOptions,
-      ...(turnAdoptionLifecycle ? { turnAdoptionLifecycle } : {}),
-    },
-  };
-}
-
-function resolveObserveOnlyDispatchResult<TDispatchResult>(
-  params: PreparedChannelTurn<TDispatchResult>,
-): TDispatchResult {
-  return (params.observeOnlyDispatchResult ?? {
-    queuedFinal: false,
-    counts: EMPTY_CHANNEL_TURN_DISPATCH_COUNTS,
-  }) as TDispatchResult;
-}
-
-function isSystemChannelTurn(ctx: FinalizedMsgContext): boolean {
-  return (
-    ctx.Provider === "heartbeat" || ctx.Provider === "cron-event" || ctx.Provider === "exec-event"
-  );
-}
-
-function maybeWarnZeroCountVisibleDispatch<TDispatchResult>(
-  params: Pick<
-    PreparedChannelTurn<TDispatchResult>,
-    "admission" | "channel" | "ctxPayload" | "messageId" | "routeSessionKey"
-  > & {
-    dispatchResult: TDispatchResult;
-    log?: (event: ChannelTurnLogEvent) => void;
-  },
-): void {
-  if (params.admission?.kind === "observeOnly" || isSystemChannelTurn(params.ctxPayload)) {
-    return;
-  }
-  const dispatchResult = params.dispatchResult as ChannelTurnDispatchResultLike;
-  // Suppress the silent-drop warning using the canonical visible-delivery signal, which
-  // includes observedReplyDelivery and other non-count delivery paths. A partial count-only
-  // check would falsely flag observed-path deliveries (queuedFinal=false, zero counts) as drops.
-  if (hasVisibleChannelTurnDispatch(dispatchResult, NO_ADDITIONAL_DELIVERY_SIGNALS)) {
-    return;
-  }
-  log.warn(
-    `visible channel turn dispatched with no queued reply payloads: channel=${params.channel} ` +
-      `messageId=${params.messageId ?? "unknown"} sessionKey=${
-        params.ctxPayload.SessionKey ?? params.routeSessionKey
-      }`,
-  );
-  emit({
-    ...params,
-    event: {
-      stage: "dispatch",
-      event: "warning",
-      messageId: params.messageId,
-      sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-      admission: params.admission?.kind ?? "dispatch",
-      reason: "zero-count-visible-dispatch",
-    },
-  });
-}
-
-function isExplicitlyNonVisibleChannelDelivery(result: unknown): boolean {
-  return (
-    typeof result === "object" &&
-    result !== null &&
-    !Array.isArray(result) &&
-    (result as { visibleReplySent?: unknown }).visibleReplySent === false
-  );
-}
-
-function markChannelDeliveryErrorVisible(error: unknown): unknown {
-  if (typeof error === "object" && error !== null && !Array.isArray(error)) {
-    try {
-      Object.assign(error, { sentBeforeError: true, visibleReplySent: true });
-      return error;
-    } catch {
-      // Fall back to a wrapper when a platform error object is non-extensible.
-    }
-  }
-  const visibleError = new Error("visible channel reply delivery failed", { cause: error });
-  Object.assign(visibleError, { sentBeforeError: true, visibleReplySent: true });
-  return visibleError;
-}
-
-async function runChannelDeliveryObserver(params: {
-  onDelivered: ChannelEventDeliveryAdapter["onDelivered"] | undefined;
-  payload: ReplyPayload;
-  info: Parameters<NonNullable<ChannelEventDeliveryAdapter["onDelivered"]>>[1];
-  result: Parameters<NonNullable<ChannelEventDeliveryAdapter["onDelivered"]>>[2];
-}): Promise<void> {
-  if (!params.onDelivered) {
-    return;
-  }
-  try {
-    await params.onDelivered(params.payload, params.info, params.result);
-  } catch (error: unknown) {
-    throw isExplicitlyNonVisibleChannelDelivery(params.result)
-      ? error
-      : markChannelDeliveryErrorVisible(error);
-  }
-}
-
-function resolveBotLoopProtectionDrop<TDispatchResult>(
-  params: PreparedChannelTurn<TDispatchResult>,
-): ChannelTurnResult<TDispatchResult> | undefined {
-  if (!params.botLoopProtection) {
-    return undefined;
-  }
-  const botLoopResult = recordChannelBotPairLoopAndCheckSuppression(params.botLoopProtection);
-  if (!botLoopResult.suppressed) {
-    return undefined;
-  }
-  const admission: ChannelTurnAdmission = { kind: "drop", reason: "bot-loop-protection" };
-  emit({
-    ...params,
-    event: {
-      stage: "authorize",
-      event: "drop",
-      messageId: params.messageId,
-      sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-      admission: admission.kind,
-      reason: admission.reason,
-    },
-  });
-  return {
-    admission,
-    dispatched: false,
-    ctxPayload: params.ctxPayload,
-    routeSessionKey: params.routeSessionKey,
-  };
-}
-
-type AssembledChannelTurnWithBotLoopProtection = AssembledChannelTurn & {
-  botLoopProtection: NonNullable<AssembledChannelTurn["botLoopProtection"]>;
-};
-
-type AssembledChannelTurnWithoutBotLoopProtection = Omit<
-  AssembledChannelTurn,
-  "botLoopProtection"
-> & {
-  botLoopProtection?: undefined;
-};
-
-export function dispatchAssembledChannelTurn(
-  params: AssembledChannelTurnWithBotLoopProtection,
-): Promise<ChannelTurnResult>;
-export function dispatchAssembledChannelTurn(
-  params: AssembledChannelTurnWithoutBotLoopProtection,
-): Promise<DispatchedChannelTurnResult>;
-export function dispatchAssembledChannelTurn(
-  params: AssembledChannelTurn,
-): Promise<ChannelTurnResult>;
-export async function dispatchAssembledChannelTurn(
-  params: AssembledChannelTurn,
-): Promise<ChannelTurnResult> {
-  const replyPipeline = resolveAssembledReplyPipeline(params);
-  return await runPreparedChannelTurnCore(
-    {
-      channel: params.channel,
-      accountId: params.accountId,
-      routeSessionKey: params.routeSessionKey,
-      storePath: params.storePath,
-      ctxPayload: params.ctxPayload,
-      recordInboundSession: params.recordInboundSession,
-      afterRecord: params.afterRecord,
-      record: params.record,
-      history: params.history,
-      admission: params.admission,
-      botLoopProtection: params.botLoopProtection,
-      log: params.log,
-      messageId: params.messageId,
-      runDispatch: async () =>
-        await runWithSessionInitConflictRetry(
-          () =>
-            params.dispatchReplyWithBufferedBlockDispatcher({
-              ctx: params.ctxPayload,
-              cfg: params.cfg,
-              dispatcherOptions: {
-                ...replyPipeline.dispatcherOptions,
-                deliver: async (payload: ReplyPayload, info) => {
-                  const preparedPayload = params.delivery.preparePayload
-                    ? await params.delivery.preparePayload(payload, info)
-                    : payload;
-                  const durableOptions =
-                    typeof params.delivery.durable === "function"
-                      ? await params.delivery.durable(preparedPayload, info)
-                      : params.delivery.durable;
-                  if (durableOptions) {
-                    const durable = await deliverInboundReplyWithMessageSendContext({
-                      cfg: params.cfg,
-                      channel: params.channel,
-                      accountId: params.accountId,
-                      agentId: params.agentId,
-                      ctxPayload: params.ctxPayload,
-                      payload: preparedPayload,
-                      info,
-                      ...durableOptions,
-                    });
-                    throwIfDurableInboundReplyDeliveryFailed(durable);
-                    if (isDurableInboundReplyDeliveryHandled(durable)) {
-                      await runChannelDeliveryObserver({
-                        onDelivered: params.delivery.onDelivered,
-                        payload: preparedPayload,
-                        info,
-                        result: durable.delivery,
-                      });
-                      return durable.delivery;
-                    }
-                  }
-                  const result = await params.delivery.deliver(preparedPayload, info);
-                  await runChannelDeliveryObserver({
-                    onDelivered: params.delivery.onDelivered,
-                    payload: preparedPayload,
-                    info,
-                    result,
-                  });
-                  return result;
-                },
-                onError: params.delivery.onError,
-              },
-              toolsAllow: params.toolsAllow,
-              replyOptions: replyPipeline.replyOptions,
-              replyResolver: params.replyResolver,
-            }),
-          params.sessionInitRetry
-            ? {
-                retryDelaysMs: params.sessionInitRetry.delaysMs,
-                signal: params.sessionInitRetry.signal,
-                sleep: params.sessionInitRetry.sleep,
-              }
-            : undefined,
-        ),
-    },
-    { suppressObserveOnlyDispatch: false },
-  );
-}
-
-export const dispatchChannelInboundReply = dispatchAssembledChannelTurn;
-
-export function dispatchChannelInboundTurn(
-  plan: ChannelTurnPlan & {
-    botLoopProtection: NonNullable<ChannelTurnPlan["botLoopProtection"]>;
-  },
-): Promise<ChannelTurnResult>;
-export function dispatchChannelInboundTurn(
-  plan: Omit<ChannelTurnPlan, "botLoopProtection"> & { botLoopProtection?: undefined },
-): Promise<DispatchedChannelTurnResult>;
-export function dispatchChannelInboundTurn(plan: ChannelTurnPlan): Promise<ChannelTurnResult>;
-export async function dispatchChannelInboundTurn(
-  plan: ChannelTurnPlan,
-): Promise<ChannelTurnResult> {
-  return await dispatchAssembledChannelTurn(
-    assembleResolvedChannelTurn(plan) as AssembledChannelTurn,
-  );
-}
-
-function isPreparedChannelTurn<TDispatchResult>(
-  value: AssembledChannelTurn | PreparedChannelTurn<TDispatchResult>,
-): value is PreparedChannelTurn<TDispatchResult> & {
-  admission?: Extract<ChannelTurnAdmission, { kind: "dispatch" | "observeOnly" }>;
-} {
-  return "runDispatch" in value;
-}
-
-async function dispatchResolvedChannelTurn<TDispatchResult>(
-  params: (AssembledChannelTurn | PreparedChannelTurn<TDispatchResult>) & {
-    admission: Extract<ChannelTurnAdmission, { kind: "dispatch" | "observeOnly" }>;
-    log?: (event: ChannelTurnLogEvent) => void;
-    messageId?: string;
-  },
-): Promise<ChannelTurnResult<TDispatchResult>> {
-  if (isPreparedChannelTurn(params)) {
-    return await runPreparedChannelTurn(params);
-  }
-  return (await dispatchAssembledChannelTurn(params)) as ChannelTurnResult<TDispatchResult>;
-}
-
-async function runPreparedChannelTurnCore<
-  TDispatchResult = DispatchedChannelTurnResult["dispatchResult"],
->(
-  params: PreparedChannelTurn<TDispatchResult>,
-  options: { suppressObserveOnlyDispatch: boolean },
-): Promise<ChannelTurnResult<TDispatchResult>> {
-  const trace = createDiagnosticTraceContextFromActiveScope();
-  return await runWithDiagnosticTraceContext(trace, () =>
-    runPreparedChannelTurnCoreInTrace(params, options),
-  );
-}
-
-async function runPreparedChannelTurnCoreInTrace<
-  TDispatchResult = DispatchedChannelTurnResult["dispatchResult"],
->(
-  params: PreparedChannelTurn<TDispatchResult>,
-  options: { suppressObserveOnlyDispatch: boolean },
-): Promise<ChannelTurnResult<TDispatchResult>> {
-  const admission = params.admission ?? ({ kind: "dispatch" } as const);
-  const botLoopDrop = resolveBotLoopProtectionDrop(params);
-  if (botLoopDrop) {
-    clearPendingHistoryAfterTurn(params.history);
-    return botLoopDrop;
-  }
-  emit({
-    ...params,
-    event: {
-      stage: "record",
-      event: "start",
-      messageId: params.messageId,
-      sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-      admission: admission.kind,
-    },
-  });
-  try {
-    await params.recordInboundSession({
-      storePath: params.storePath,
-      sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-      ctx: params.ctxPayload,
-      groupResolution: params.record?.groupResolution,
-      createIfMissing: params.record?.createIfMissing,
-      updateLastRoute: params.record?.updateLastRoute,
-      onRecordError: params.record?.onRecordError ?? (() => undefined),
-      trackSessionMetaTask: params.record?.trackSessionMetaTask,
-    });
-    emit({
-      ...params,
-      event: {
-        stage: "record",
-        event: "done",
-        messageId: params.messageId,
-        sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-        admission: admission.kind,
-      },
-    });
-    await params.afterRecord?.();
-  } catch (err) {
-    emit({
-      ...params,
-      event: {
-        stage: "record",
-        event: "error",
-        messageId: params.messageId,
-        sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-        admission: admission.kind,
-        error: err,
-      },
-    });
-    try {
-      await params.onPreDispatchFailure?.(err);
-    } catch {
-      // Preserve the original session-recording error.
-    }
-    throw err;
-  }
-
-  emit({
-    ...params,
-    event: {
-      stage: "dispatch",
-      event: "start",
-      messageId: params.messageId,
-      sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-      admission: admission.kind,
-    },
-  });
-  let dispatchResult: TDispatchResult;
-  try {
-    dispatchResult =
-      options.suppressObserveOnlyDispatch && admission.kind === "observeOnly"
-        ? resolveObserveOnlyDispatchResult(params)
-        : await params.runDispatch();
-    maybeWarnZeroCountVisibleDispatch({
-      ...params,
-      admission,
-      dispatchResult,
-    });
-  } catch (err) {
-    emit({
-      ...params,
-      event: {
-        stage: "dispatch",
-        event: "error",
-        messageId: params.messageId,
-        sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-        admission: admission.kind,
-        error: err,
-      },
-    });
-    throw err;
-  }
-  emit({
-    ...params,
-    event: {
-      stage: "dispatch",
-      event: "done",
-      messageId: params.messageId,
-      sessionKey: params.ctxPayload.SessionKey ?? params.routeSessionKey,
-      admission: admission.kind,
-    },
-  });
-  clearPendingHistoryAfterTurn(params.history);
-
-  return {
-    admission,
-    dispatched: true,
-    ctxPayload: params.ctxPayload,
-    routeSessionKey: params.routeSessionKey,
-    dispatchResult,
-  };
-}
-
-type PreparedChannelTurnWithBotLoopProtection<TDispatchResult> =
-  PreparedChannelTurn<TDispatchResult> & {
-    botLoopProtection: NonNullable<PreparedChannelTurn<TDispatchResult>["botLoopProtection"]>;
-  };
-
-type PreparedChannelTurnWithoutBotLoopProtection<TDispatchResult> = Omit<
-  PreparedChannelTurn<TDispatchResult>,
-  "botLoopProtection"
-> & {
-  botLoopProtection?: undefined;
-};
-
-function runPreparedChannelTurn<TDispatchResult = DispatchedChannelTurnResult["dispatchResult"]>(
-  params: PreparedChannelTurnWithBotLoopProtection<TDispatchResult>,
-): Promise<ChannelTurnResult<TDispatchResult>>;
-function runPreparedChannelTurn<TDispatchResult = DispatchedChannelTurnResult["dispatchResult"]>(
-  params: PreparedChannelTurnWithoutBotLoopProtection<TDispatchResult>,
-): Promise<DispatchedChannelTurnResult<TDispatchResult>>;
-function runPreparedChannelTurn<TDispatchResult = DispatchedChannelTurnResult["dispatchResult"]>(
-  params: PreparedChannelTurn<TDispatchResult>,
-): Promise<ChannelTurnResult<TDispatchResult>>;
-async function runPreparedChannelTurn<
-  TDispatchResult = DispatchedChannelTurnResult["dispatchResult"],
->(params: PreparedChannelTurn<TDispatchResult>): Promise<ChannelTurnResult<TDispatchResult>> {
-  return await runPreparedChannelTurnCore(params, { suppressObserveOnlyDispatch: true });
-}
-
-export const runPreparedInboundReply = runPreparedChannelTurn;
 
 async function runChannelTurn<
   TRaw,
@@ -772,19 +247,15 @@ async function runChannelTurn<
   const admission = resolved.admission ?? preflightAdmission ?? ({ kind: "dispatch" } as const);
   let result: ChannelTurnResult<TDispatchResult>;
   try {
-    // Prepared runDispatch was assembled earlier and ignores late options.
-    const dispatchResult = await dispatchResolvedChannelTurn(
+    const dispatchResult = (
       "runDispatch" in resolved
-        ? {
+        ? await runPreparedInboundReply({
             ...resolved,
-            ...(admission.kind === "observeOnly"
-              ? { delivery: createNoopChannelEventDeliveryAdapter() }
-              : {}),
             admission,
             log: params.log,
             messageId: input.id,
-          }
-        : {
+          })
+        : await dispatchAssembledChannelTurn({
             ...resolved,
             ...(admission.kind === "observeOnly"
               ? { delivery: createNoopChannelEventDeliveryAdapter() }
@@ -795,8 +266,8 @@ async function runChannelTurn<
             ...(params.turnAdoptionLifecycle
               ? { turnAdoptionLifecycle: params.turnAdoptionLifecycle }
               : {}),
-          },
-    );
+          })
+    ) as ChannelTurnResult<TDispatchResult>;
     result = dispatchResult.dispatched ? { ...dispatchResult, admission } : dispatchResult;
   } catch (err) {
     const failedResult: ChannelTurnResult<TDispatchResult> = {
@@ -857,4 +328,3 @@ async function runChannelTurn<
 }
 
 export const runChannelInboundEvent = runChannelTurn;
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
